requirements.txt
streamlit
requests
freeimagehost